boards: adi: Add JLink support to max32670evkit
To be able to run CI/CD tests on max32670evkit JLink support is
supposed to be added.
Normally max32670evkit has max32625pico on it. Therefore, the user can
only flash the board using max32625pico that supports OpenOCD CMSIS-DAP.
However, ADI CI/CD environment works with J-Link. Therefore, ADI CI/CD
environment uses a modified version of max32670evkit so that it can be
flashed using J-Link.
Since J-Link is not an option for a customer. This PR should not be
pushed to the upstream.

Signed-off-by: Anıl Kara <anil.kara@analog.com>