ci: workflows: Remove existing access tokens
Most of the workflows generates temporary access token and add this
token to the global .gitconfig file. Normally this file is removed;
therefore, all generated ids do not accumulate in the .gitconfig file.
However, in some cases the file is not cleaned and generated tokens
starts to accumulate in .gitconfig file. To remove already existing
tokens from .gitconfig file this workflow step is added.

Signed-off-by: Anıl Kara <anil.kara@analog.com>