bluetooth: services: cgms: Add CGM Service
This implementation of the Continuous Glucose Monitoring Profile
includes all functionality marked as 'required', including the
ability to query records using the 'greater- or-equal' filter.
The service requires three callback functions to operate. These
callbacks allow the user to customize what data is stored during
each interval, as well as how records are stored during the lifetime
of the device.

For more information on BLE Continuous Glucose Monitoring Profile:
https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/CGMS_v1.0.2/out/en/index-en.html

Signed-off-by: Sean Kyer <Sean.Kyer@analog.com>