boards: adi: Fix MAX78002EVKIT JLink issue
This commit changes device parameter of JLink to MAX32690 for
MAX78002EVKIT to unblock CI tests of it.

We need to use JLink in CI setup because there are more than one board
in same CI setup. Twister uses JLink serial number to identify the board.
We can not select board with OpenOCD because it does not have serial
number.

JLink does not currently support MAX78002 but MAX32690 can be used
because the flash algorithm is the same.

Signed-off-by: Furkan Akkiz <hasanfurkan.akkiz@analog.com>