manifest: Tie packetcraft-adi to specific commit
The packetcraft-adi module in the west.yml currently points at branch
'v0.0.4+a7c6c73f.20241219-libs', whereas all the other modules point
at specific commits. With Zephyr in CFS, this causes 'west update' to
fail if there is no access to adi-innersource, as it attempts to fetch
the latest commit on that branch.

That fetch does not happen when a commit hash is used instead and that
hash already matches the checked-out module, as already done for the
trusted-firmware-m and tf-m-tests modules from adi-innersource.

Signed-off-by: Andy Koppe <andy.koppe@analog.com>